% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplotR-package.R
\docType{package}
\name{forestplotR-package}
\alias{forestplotR}
\alias{forestplotR-package}
\title{forestplotR: Generates Forest Plots}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This packages takes input from a dataframe/tibble and outputs a forest plot aligned with tables to provide additional information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/forestplotR}
  \item Report bugs at \url{https://github.com/mglev1n/forestplotR/issues}
}

}
\keyword{internal}
