% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_res_example.R
\docType{data}
\name{mr_res_example}
\alias{mr_res_example}
\title{mr_res_example}
\format{
A data frame with 35 rows and 16 variables:
\describe{
  \item{\code{id.exposure}}{character COLUMN_DESCRIPTION}
  \item{\code{id.outcome}}{character COLUMN_DESCRIPTION}
  \item{\code{outcome}}{character COLUMN_DESCRIPTION}
  \item{\code{exposure}}{character COLUMN_DESCRIPTION}
  \item{\code{method}}{character COLUMN_DESCRIPTION}
  \item{\code{nsnp}}{double COLUMN_DESCRIPTION}
  \item{\code{b}}{double COLUMN_DESCRIPTION}
  \item{\code{se}}{double COLUMN_DESCRIPTION}
  \item{\code{pval}}{double COLUMN_DESCRIPTION}
  \item{\code{p_fdr}}{double COLUMN_DESCRIPTION}
  \item{\code{p_bonferroni}}{double COLUMN_DESCRIPTION}
  \item{\code{lo_ci}}{double COLUMN_DESCRIPTION}
  \item{\code{up_ci}}{double COLUMN_DESCRIPTION}
  \item{\code{or}}{double COLUMN_DESCRIPTION}
  \item{\code{or_lci95}}{double COLUMN_DESCRIPTION}
  \item{\code{or_uci95}}{double COLUMN_DESCRIPTION}
}
}
\usage{
mr_res_example
}
\description{
Example dataset containing typical output from the \code{mr} function of the \code{TwoSampleMR} package
}
\details{
DETAILS
}
\keyword{datasets}
