% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Generate forest plot with aligned data table}
\usage{
plot_forest(
  df,
  cols_left,
  ci_lower = ci_lower,
  ci_upper = ci_upper,
  ci_signif = 2,
  color = NULL,
  effect = effect,
  effect_signif = 2,
  p_value = p_value,
  size = 2,
  width_ratio = c(1, 1, 1),
  xmin = 0.5,
  xmax = 2,
  hline = 1,
  style = "OR",
  font_size = 12
)
}
\arguments{
\item{df}{a data frame}

\item{cols_left}{Names of columns to be included in a table that will be displayed to the left of the forest plot, passed within the \code{vars()} function. These may be descriptors that differentiate the values displayed in the forest plot.}

\item{ci_lower}{Name of column containing values of lower limit of confidence interval}

\item{ci_upper}{Name of column containing values of upper limit of confidence interval}

\item{ci_signif}{An integer passed to \code{signif} used to round the confidence interval to the specified number of significant digits}

\item{color}{Name of column that should be used to color points in forest plot. The default value is \code{NULL}}

\item{effect}{Name of column containing effect estimates (eg. Odds ratio or Beta)}

\item{effect_signif}{An integer passed to \code{signif} used to round the effect estimate to the specified number of significant digits}

\item{p_value}{Name of column containing P-value}

\item{size}{Integer or name of column used to scale points on forest plot}

\item{width_ratio}{Vector of integers representing relative widths of pieces of final plot. There should be one integer for each column passed to cols_left, the forest plot, and the results table.}

\item{xmin}{Integer representing the lower limit of the forest plot. If greater than the lowest value of the lower confidence interval, forest plot values will be truncated with an arrow. The lower confidence interval column can be passed in the following code to ensure no truncation occurs \code{xmin = min(c(1,.$or_lci95))}.}

\item{xmax}{Integer representing the lower limit of the forest plot. If less than the greatest value of the upper confidence interval, forest plot values will be truncated with an arrow. The upper confidence interval column can be passed in the following code to ensure no truncation occurs \code{xmin = min(c(1,.$or_lci95))}}

\item{hline}{Integer representing the location of vertical lines to be placed on the forest plot, useful for describing the origin (eg. \code{hline = 0} when plotting Odds Ratios, and \code{hline = 0} for betas)}

\item{style}{String Either "OR" or "Beta" representing the format of the effect estimate, used for setting axis style and effect-estimate column headings}

\item{font_size}{Integer representing default font size}
}
\value{
A ggplot object containing a forest plot
}
\description{
\code{plot_forest} is a function that takes a data frame that at minimum contains columns containing an effect estimate, upper and lower confidence interval bounds, and p-values. The function returns a forest plot aligned with a data table containing this information. Additional variables may be displayed in a table to the left of the plot by passing their names to the \code{cols_left} argument.
}
\examples{
plot_forest(mr_res_example, cols_left = dplyr::vars("Protein" = exposure), effect = or, effect_signif = 3, ci_lower = or_lci95, ci_upper = or_uci95, ci_signif = 3, p_value = pval, width_ratio = c(0.5, 3, 1.5), xmin = min(c(1, mr_res_example$or_lci95)), xmax = max(c(1, mr_res_example$or_uci95)), style = "OR")

}
